+++ libs/utils/CMakeLists.txt
@@ -1,12 +1,8 @@
-cmake_minimum_required(VERSION 3.22)
-
 # Define the library name
 set(LIBRARY_NAME utils)
 
 # Set project info
 project(${LIBRARY_NAME} VERSION 1.0.0 DESCRIPTION "Utility lib for TaskMaster")
-
-set(CMAKE_CXX_STANDARD 23)
 
 # Define the lib directory
 set(LIB_DIR .)
